refactored some code, added queries button
